feat: put logics
