+++ ExtremeRoles/Module/ApiHandler/PutOption.cs
@@ -8,12 +8,21 @@ namespace ExtremeRoles.Module.ApiHandler;
 
 #nullable enable
 
+public readonly record struct OptionPutRequest(int TabId, int CategoryId, int OptionId, int Selection);
+
 public sealed class PutOption : IRequestHandler
 {
 	public Action<HttpListenerContext> Request => this.requestAction;
 
 	private void requestAction(HttpListenerContext context)
 	{
+		var response = context.Response;
+		var newOptionSelection = IRequestHandler.DeserializeJson<OptionPutRequest>(context.Request);
 
+		OptionManager.Instance.Update(
+			(OptionTab)newOptionSelection.TabId,
+			newOptionSelection.CategoryId,
+			newOptionSelection.OptionId,
+			newOptionSelection.Selection);
 	}
 }
