+++ ExtremeRoles/Roles/ExtremeRoleManager.cs
@@ -281,8 +281,8 @@ public static class ExtremeRoleManager
 {
     public const int OptionOffsetPerRole = 200;
 
-	public const int RoleCategoryIdOffset = 200;
-	private const int conbRoleIdOffset = 1000;
+	// Id空間確保のため通常役職の最初のスペクル分減らす
+	public const int RoleCategoryIdOffset = -(int)ExtremeRoleId.SpecialCrew;
 
 	public static readonly IReadOnlySet<ExtremeRoleId> SpecialWinCheckRole = new HashSet<ExtremeRoleId>()
     {
@@ -443,7 +443,7 @@ public static class ExtremeRoleManager
 		=> RoleCategoryIdOffset + (int)roleId;
 
 	public static int GetCombRoleGroupId(CombinationRoleType roleId)
-		=> conbRoleIdOffset + (int)roleId;
+		=> (int)roleId;
 
 	public static void CreateCombinationRoleOptions()
     {
